Apply Fixes to Task 14 and Several Test Files
Change data type for function in 14-binary_tree_balance.c to match change made in binary_trees.h subsequently, in a more recent task.

Also delete redundant copied lines in several *-main.c test files in the tests folder. The lines were for a support function copied over from a previous file (8-main.c) when 9-main.c was created, and remained thereafter as each subsequent file was created from its predecessor.

+++ 14-binary_tree_balance.c
@@ -9,7 +9,7 @@
  * Return: Calculated balance factor of given binary tree.
  */
 
-size_t calc_balance(const binary_tree_t *tree, size_t balance)
+int calc_balance(const binary_tree_t *tree, size_t balance)
 {
 	if (!tree)
 		return (balance);
@@ -45,7 +45,7 @@ size_t calc_balance(const binary_tree_t *tree, size_t balance)
  * Return: Balance factor of given binary tree.
  */
 
-size_t binary_tree_balance(const binary_tree_t *tree)
+int binary_tree_balance(const binary_tree_t *tree)
 {
 	if (!tree)
 		return (0);
